requirements.txt
python-dotenv
azure-identity
azure-storage-queue
azure-storage-blob
azure-cosmos
azure-functions
